packaged-agent/pyproject.toml
[project]
name = "packaged-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "alihamza6682685", email = "alihamza6682685@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.11",
    "python-dotenv>=1.1.1",
]

[project.scripts]
packaged-agent = "packaged_agent.main:run"

[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"
